ir+metainterp: move InvalidLoop/SpeculativeError/INFO_* to majit-ir::optimize
Hosts the leaf-most optimizer types in majit-ir so the IntBound / OpInfo
moves that follow can reference them without re-introducing a
majit-metainterp → majit-ir dependency edge. metainterp re-exports the
same names so callers stay unchanged.
