+++ majit/majit-ir/src/lib.rs
@@ -5,6 +5,7 @@ pub mod descr_registry;
 pub mod effectinfo;
 pub mod op_descr;
 pub mod op_type_index;
+pub mod optimize;
 pub mod resoperation;
 pub mod resumecode;
 pub mod resumedata;
